Create Index of first 1 in a sorted array of 0s and 1s
Given a sorted array consisting 0’s and 1’s. The task is to find the index of first ‘1’ in the given array.

  
Your Task:  
You don't need to read input or print anything. Your task is to complete the function firstIndex() which takes the array A[] and its size N as inputs and returns the index of first 1. If 1 is not present in the array then return -1.